Update on 2019-11-20

+++ src/main/java/com/lee/msims/pojo/common/Course.java
@@ -9,13 +9,13 @@ public class Course {
     private String courseCode;
     private String faculty;
     private String name;
-    private double credit;
+    private int credit;
     private String teacher;
     private String duration;    //9.4-12.17
     private String time;        //Monday@11:00-12:45,Tuesday@9:00-10:45
     private String description;
 
-    public Course(String courseCode, String faculty, String name, double credit,
+    public Course(String courseCode, String faculty, String name, int credit,
                   String teacher, String duration, String time, String description) {
         this.courseCode = courseCode;
         this.faculty = faculty;
